requirements.txt
pdfminer
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master
